requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlmodel==0.0.14
asyncpg==0.29.0
alembic==1.12.1
pgvector==0.2.4

# AI and ML
langchain==0.1.0
langchain-openai==0.0.2
openai==1.3.7
sentence-transformers==2.2.2
torch==2.1.1

# HTTP client
httpx==0.25.2

# Background tasks
arq==0.25.0
redis==5.0.1

# Authentication and security
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring and logging
prometheus-fastapi-instrumentator==6.1.0
structlog==23.2.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Supabase
supabase==2.3.0
postgrest==0.13.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1
